+++ package.json
@@ -6,6 +6,7 @@
     "start": "node server.js"
   },
   "dependencies": {
+    "axios": "^1.10.0",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
